Refactor: Extract new mod `utils/rgb.rs`
