+++ graphicsmagick/src/utils/rgb.rs
@@ -1,0 +1,23 @@
+use std::os::raw::{c_double, c_uint};
+
+/// For [`max_rgb`] to return max RGB value.
+pub trait MaxRGB {
+    fn max_rgb() -> Self;
+}
+
+impl MaxRGB for c_uint {
+    fn max_rgb() -> Self {
+        graphicsmagick_sys::MaxRGB
+    }
+}
+
+impl MaxRGB for c_double {
+    fn max_rgb() -> Self {
+        graphicsmagick_sys::MaxRGBDouble
+    }
+}
+
+/// Wrapper of `graphicsmagick_sys::MaxRGB` and `graphicsmagick_sys::MaxRGBDouble`.
+pub fn max_rgb<T: MaxRGB>() -> T {
+    <T>::max_rgb()
+}
